typedarray: Use a Cell for handling interior mutability.
The typed array wrapper has a "computed" option that uses to cache its data.

That means that you need a mutable reference to the typed array in order to read
it, which is not great.

In particular, it's plain impossible when the typed array is provided as a
member of a struct via WebIDL.

Use a `Cell` and expose to_vec and as_slice in non-mutable references.