Create a second set of borrowing wrappers
Our current wrappers held in rust::wrappers accept the Handles by
consume/Copy. This was introduced in #393 to make it feasible to migrate
the API calls on the Servo side in finite time. This, in particular,
requires unsafe code in the implementation of MutableHandle.

This change introduces a second set of wrappers, which will now borrow
MutableHandles instead of relying on impl Copy for MutableHandle. The
Servo code should gradually move to the new wrappers,
rust::jsapi_wrapped, so that we can eventually remove rust::wrappers.